Updated codegen config/renamed file
