+++ codegen.ts
@@ -5,22 +5,24 @@ const config: CodegenConfig = {
   documents: ['src/gql/apolloGQL/**/*.ts'],
   overwrite: true,
   generates: {
-    './src/gql/gen/graphql-types.ts': {
+    './src/gql/gen/': {
+      preset: 'client',
+      presetConfig: {
+        fragmentMasking: false
+      }
+    },
+    './src/gql/gen/graphql.ts': {
       plugins: [
         'typescript',
         'typescript-operations',
         'typescript-react-apollo'
       ],
       config: {
-        withHooks: true
-      }
-    },
-    './src/gql/gen/': {
-      preset: 'client',
-      presetConfig: {
-        fragmentMasking: false
-      },
-      config: {
+        withHooks: true,
+        nonOptionalTypename: true,
+        namingConvention: {
+          enumValues: 'change-case-all#upperCase#snakeCase'
+        },
         scalars: {
           // old codegen mappings from global.d.ts
           // maintain until we add better scalar mapping with graphql-codegen
@@ -30,10 +32,6 @@ const config: CodegenConfig = {
           Time: 'Time',
           UUID: 'UUID',
           Upload: 'Upload'
-        },
-        nonOptionalTypename: true,
-        namingConvention: {
-          enumValues: 'change-case-all#upperCase#snakeCase'
         }
       }
     }
