+++ src/views/ModelPlan/TaskList/Payment/FundingSource/index.tsx
@@ -15,11 +15,12 @@ import {
   TextInput
 } from '@trussworks/react-uswds';
 import { Field, Form, Formik, FormikProps } from 'formik';
-import { GetFundingQuery, PayType } from 'gql/gen/graphql';
 import {
+  GetFundingQuery,
+  PayType,
   useGetFundingQuery,
   useUpdatePaymentsMutation
-} from 'gql/gen/graphql-types';
+} from 'gql/gen/graphql';
 
 import AddNote from 'components/AddNote';
 import AskAQuestion from 'components/AskAQuestion';
