+++ src/views/Unfollow/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useHistory, useLocation } from 'react-router-dom';
 import { useMutation } from '@apollo/client';
-import { useGetBasicsQuery } from 'gql/gen/graphql-types';
+import { useGetBasicsQuery } from 'gql/gen/graphql';
 
 import PageLoading from 'components/PageLoading';
 import Alert from 'components/shared/Alert';
